+++ runtime/os.go
@@ -1,6 +1,8 @@
-package util
+package runtime
 
-import "runtime"
+import (
+	r "runtime"
+)
 
 var (
 	// IsWindows は実行環境がWindowsであるかどうかを返す、テスト時は値を後で設定可能
@@ -8,5 +10,5 @@ var (
 )
 
 func init() {
-	IsWindows = runtime.GOOS == "windows"
+	IsWindows = r.GOOS == "windows"
 }
